requirement.txt
cffi
cryptography
mysql-connector-python
prettytable
pycparser
pycryptodome
wcwidth